+++ playground/vitest.config.e2e.ts
@@ -1,9 +1,7 @@
 import { resolve } from 'node:path'
-import { defaultExclude, defineConfig } from 'vitest/config'
+import { defineConfig } from 'vitest/config'
 
 const timeout = process.env.PWDEBUG ? Infinity : process.env.CI ? 20_000 : 5_000
-
-const isBelowNode20 = +process.versions.node.split('.')[0] < 20
 
 export default defineConfig({
   resolve: {
@@ -14,9 +12,6 @@ export default defineConfig({
   test: {
     pool: 'forks',
     include: ['./playground/**/*.spec.[tj]s'],
-    exclude: isBelowNode20
-      ? ['**/__tests__/oxc/**', ...defaultExclude] // plugin-oxc only supports node >= 20
-      : defaultExclude,
     setupFiles: ['./playground/vitestSetup.ts'],
     globalSetup: ['./playground/vitestGlobalSetup.ts'],
     testTimeout: timeout,